bracket-push: rename to matching-brackets
RE: https://github.com/exercism/problem-specifications/pull/1501
