+++ config.json
@@ -798,7 +798,7 @@
       ]
     },
     {
-      "slug": "bracket-push",
+      "slug": "matching-brackets",
       "uuid": "45229a7c-6703-4240-8287-16645881a043",
       "core": false,
       "unlocked_by": "isogram",
